requirements.txt
flask
google-auth
google-auth-oauthlib
google-api-python-client
PyMuPDF
